# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: General Supplies Procurement Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer-pos:
    get:
      operationId: listCustomerPos
      tags: [customer-pos]
      summary: List all customer purchase orders
      responses:
        "200":
          description: List of customer POs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerPo"
    post:
      operationId: createCustomerPo
      tags: [customer-pos]
      summary: Create a customer purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerPoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPo"
  /customer-pos/{id}:
    get:
      operationId: getCustomerPo
      tags: [customer-pos]
      summary: Get a customer PO by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer PO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPo"
        "404":
          description: Not found
    patch:
      operationId: updateCustomerPo
      tags: [customer-pos]
      summary: Update a customer PO
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerPoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPo"

  # ─── SUPPLIER POs ─────────────────────────────────────────────────────────────
  /supplier-pos:
    get:
      operationId: listSupplierPos
      tags: [supplier-pos]
      summary: List all supplier purchase orders
      responses:
        "200":
          description: List of supplier POs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierPo"
    post:
      operationId: createSupplierPo
      tags: [supplier-pos]
      summary: Create a supplier purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierPoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPo"
  /supplier-pos/{id}:
    get:
      operationId: getSupplierPo
      tags: [supplier-pos]
      summary: Get a supplier PO by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier PO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPo"
        "404":
          description: Not found
    patch:
      operationId: updateSupplierPo
      tags: [supplier-pos]
      summary: Update a supplier PO
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierPoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPo"

  # ─── DASHBOARD ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CustomerPo:
      type: object
      required: [id, customerId, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        quotationId:
          type: ["integer", "null"]
        poNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, processing, fulfilled, cancelled]
        totalAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CustomerPoInput:
      type: object
      required: [customerId]
      properties:
        customerId:
          type: integer
        quotationId:
          type: integer
        poNumber:
          type: string
        status:
          type: string
          enum: [received, processing, fulfilled, cancelled]
        totalAmount:
          type: number
        notes:
          type: string
    CustomerPoUpdate:
      type: object
      properties:
        poNumber:
          type: string
        status:
          type: string
          enum: [received, processing, fulfilled, cancelled]
        totalAmount:
          type: number
        notes:
          type: string

    # ─── SUPPLIER PO ───────────────────────────────────────────────────────────
    SupplierPo:
      type: object
      required: [id, supplierId, status, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: ["string", "null"]
        customerPoId:
          type: ["integer", "null"]
        poNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, confirmed, delivered, cancelled]
        totalAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SupplierPoInput:
      type: object
      required: [supplierId]
      properties:
        supplierId:
          type: integer
        customerPoId:
          type: integer
        poNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, confirmed, delivered, cancelled]
        totalAmount:
          type: number
        notes:
          type: string
    SupplierPoUpdate:
      type: object
      properties:
        poNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, confirmed, delivered, cancelled]
        totalAmount:
          type: number
        notes:
          type: string

    # ─── DASHBOARD ─────────────────────────────────────────────────────────────
...