# lib/api-zod/src/generated/types/customerPoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * General Supplies Procurement Management System
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerPoUpdateStatus } from './customerPoUpdateStatus';

export interface CustomerPoUpdate {
  poNumber?: string;
  status?: CustomerPoUpdateStatus;
  totalAmount?: number;
  notes?: string;
}