# lib/api-zod/src/generated/types/supplierPoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * General Supplies Procurement Management System
 * OpenAPI spec version: 0.1.0
 */
import type { SupplierPoUpdateStatus } from './supplierPoUpdateStatus';

export interface SupplierPoUpdate {
  poNumber?: string;
  status?: SupplierPoUpdateStatus;
  totalAmount?: number;
  notes?: string;
}